LayoutTests/platform/qt-5.0-wk1/http/tests/security/contentSecurityPolicy/object-src-none-allowed-expected.txt

CONSOLE MESSAGE: Refused to load plugin data from '' because it violates the following Content Security Policy directive: "object-src 'none'".

CONSOLE MESSAGE: Refused to load plugin data from '' because it violates the following Content Security Policy directive: "object-src 'none'".



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS!
